% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{r_logo_example}
\alias{r_logo_example}
\title{Load a 2D data set with the shape of the R logo}
\usage{
r_logo_example()
}
\value{
A data frame with the x and y coordinates of the R logo
}
\description{
This function is a helper function to load the R logo data included under
the \verb{inst/} directory.
}
\examples{
# Load the R logo data
data <- r_logo_example()

# Plot the R logo
plot(data$x, data$y, pch = 16, main = "R Logo", xlab = '', ylab = '')
}
\references{
Staudenmayer, J. (2007). Hidden Images in R. Retrieved from
\url{https://www4.stat.ncsu.edu/~stefansk/NSF_Supported/Hidden_Images/000_R_Programs/John_Staudenmayer/logo.txt}
}
