% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{r_logo_image_data}
\alias{r_logo_image_data}
\title{R Logo Pixel Data}
\format{
A data frame with 2,000 observations and 2 variables describing the
x and y coordinates of the R logo.
}
\usage{
r_logo_image_data
}
\description{
2D data set with the shape of the R Logo in x and y coordinate pairings.
}
\examples{
# Load the R logo data
data("r_logo_image_data", package = "surreal")

# Plot the R logo
plot(r_logo_image_data$x, r_logo_image_data$y, pch = 16, main = "R Logo", xlab = '', ylab = '')
}
\references{
Staudenmayer, J. (2007). Hidden Images in R. Retrieved from
\url{https://www4.stat.ncsu.edu/~stefansk/NSF_Supported/Hidden_Images/000_R_Programs/John_Staudenmayer/logo.txt}
}
\keyword{datasets}
