% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_packages}
\alias{require_packages}
\title{Verify suggested packages are available}
\usage{
require_packages(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Stops with an error message if any of the required packages are missing.
Otherwise, returns \code{TRUE} invisibly.
}
\description{
Checks if the required packages are available. If not, an error message is
thrown.
}
\keyword{internal}
