% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surreal-package.R
\docType{package}
\name{surreal-package}
\alias{surreal-package}
\title{surreal: Generate Datasets with Hidden Images in Residual Plots}
\description{
Implements the "Residual (Sur)Realism" algorithm described by Stefanski (2007) \doi{10.1198/000313007X190079} to generate datasets that reveal hidden images or messages in their residual plots. It offers both predefined datasets and tools to embed custom text or images into residual structures. Allowing users to create intriguing visual demonstrations for teaching model diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/surreal}
  \item \url{https://r-pkg.thecoatlessprofessor.com/surreal/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/surreal/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
