% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surreal-text.R
\name{temporary_text_plot}
\alias{temporary_text_plot}
\title{Create a Temporary Text Plot}
\usage{
temporary_text_plot(text, cex = 4)
}
\arguments{
\item{text}{A character string to be plotted.}

\item{cex}{A numeric value specifying the relative size of the text. Default is 4.}
}
\value{
A character string containing the path to the temporary image file.
}
\description{
This function creates a temporary bitmap image file containing a plot of the given text.
}
\examples{
temp_file <- temporary_text_plot("Hello, World!")
}
\keyword{internal}
