% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surreal.R
\name{transform_data}
\alias{transform_data}
\title{Transform Data for X and Y Finding Algorithm}
\usage{
transform_data(x, y, n_add_points = 40, verbose = FALSE)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{n_add_points}{Integer. Number of points to add on each side of the frame. Default is 40.}

\item{verbose}{Logical. If TRUE, prints optimization progress. Default is FALSE.}
}
\value{
A matrix with two columns representing the transformed x and y coordinates.
}
\description{
This function transforms the input data by adding points around the original data
to create a frame. It uses an optimization process to find the best alpha parameter
for point distribution.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
transformed_data <- transform_data(x, y)

}
