% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surreal-package.R
\docType{package}
\name{surreal-package}
\alias{surreal-package}
\title{surreal: Generate Data with Hidden Images in Residual Plots}
\description{
Generate data sets so that the plot of residuals versus predicted values from the least squares fit of the correct model reveals a hidden image or message.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/surreal}
  \item \url{https://r-pkg.thecoatlessprofessor.com/surreal/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/surreal/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
